feat(lazy-init): deep freeze cached objects by default
It makes sense to deep freeze cached objects by default since there is only one object and changes
made to it will apply everywhere the same object scheme is used, which is most likely unintended, or
if it was intended, then the places where a mutable object is needed should not enable caching it.

BREAKING CHANGE:
- Option `{ freeze: true }` now deep freezes the object instead of shallow, to change
this the environment variable `LAZY_INIT_FREEZE_MODE` can be set to either `shallow` (shallow
freeze) or `none` to never freeze objects.
- Option `{ cache: true }` will now enable freezing by
default, to disable this, option `{ freeze: false }` must be explicitly set.