refactor(lazy-init): remove "lazy-init/cache" export
Now that the "lzc" method has been added to the base export I have decided to drop this way of
caching by default. It adds unnecessary complexity to the build for a one letter tradeoff.

BREAKING CHANGE: The path "lazy-init/cache" no longer exists. It should be replaced with `import {
lzc } from "lazy-init"`.